remove ssl_trusted_certificate
according to
https://community.letsencrypt.org/t/will-does-the-letsencrypt-client-create-a-cert-chain-usable-with-ocsp-stapling/2072/18

"    For the OCSP stapling to work, the certificate of the server
certificate issuer should be known. If the ssl_certificate file does not
contain intermediate certificates, the certificate of the server
certificate issuer should be present in the ssl_trusted_certificate
file.

So we don't need to specify ssl_trusted_certificate, because Let's
Encrypt creates fullchain.pem containing all certificates and we use
that for ssl_certificate already."